airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-apache-spark
pyspark==3.5.1
meteostat